@@ -9,12 +9,12 @@ plugins {
 
 android {
   namespace = "app.kaito_dogi.mybrary.feature.mybookdetail"
-  compileSdk = 34
+  compileSdk = libs.versions.compileSdk.get().toInt()
 
   resourcePrefix = "mybookdetail_"
 
   defaultConfig {
-    minSdk = 24
+    minSdk = libs.versions.minSdk.get().toInt()
 
     testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     consumerProguardFiles("consumer-rules.pro")
@@ -36,7 +36,7 @@ android {
   }
 
   kotlinOptions {
-    jvmTarget = "17"
+    jvmTarget = libs.versions.jvmTarget.get()
   }
 
   buildFeatures {
@@ -49,6 +49,7 @@ android {
 }
 
 dependencies {
+  implementation(project(":core:common"))
   implementation(project(":core:designsystem"))
   implementation(project(":core:domain"))
   implementation(project(":core:ui"))
