@@ -4,7 +4,7 @@ import app.kaito_dogi.mybrary.core.domain.model.Memo
 import app.kaito_dogi.mybrary.core.domain.model.MemoId
 import app.kaito_dogi.mybrary.core.domain.model.MyBookId
 import app.kaito_dogi.mybrary.core.domain.repository.MemoRepository
-import java.util.Date
+import java.time.LocalDateTime
 import javax.inject.Inject
 
 internal class MemoRepositoryImpl @Inject constructor() : MemoRepository {
@@ -18,10 +18,10 @@ internal class MemoRepositoryImpl @Inject constructor() : MemoRepository {
     return Memo(
       id = MemoId(0L),
       myBookId = MyBookId(0L),
-      content = "Memo",
+      content = "content",
       fromPage = null,
       toPage = null,
-      createdAt = Date(),
+      createdAt = LocalDateTime.now(),
       isPosted = false,
       postedAt = null,
       likeCount = null,
