@@ -1,15 +1,20 @@
 package app.kaito_dogi.mybrary.core.domain.model
 
-import java.util.Date
+import app.kaito_dogi.mybrary.core.common.serializer.LocalDateTimeSerializer
+import java.time.LocalDateTime
+import kotlinx.serialization.Serializable
 
+@Serializable
 data class Memo(
   val id: MemoId,
   val myBookId: MyBookId,
   val content: String,
   val fromPage: Int?,
   val toPage: Int?,
-  val createdAt: Date,
+  @Serializable(with = LocalDateTimeSerializer::class)
+  val createdAt: LocalDateTime,
   val isPosted: Boolean,
-  val postedAt: Date?,
+  @Serializable(with = LocalDateTimeSerializer::class)
+  val postedAt: LocalDateTime?,
   val likeCount: Int?,
 )
