@@ -3,5 +3,6 @@ package app.kaito_dogi.mybrary.core.domain.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-@JvmInline
-value class MyBookId(val value: Long)
+data class MyBookId(
+  val value: Long,
+)
