@@ -1,4 +1,8 @@
 package app.kaito_dogi.mybrary.core.domain.model
 
-@JvmInline
-value class UserId(val value: Long)
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UserId(
+  val value: Long,
+)
